requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.2
djangorestframework==3.14.0
flake8==5.0.4
gunicorn==20.1.0
h11==0.14.0
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
